% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{data}
\alias{data}
\title{Example data set}
\description{
Example data set
}
\references{
\url{www.impact-initiatives.org}
}
\author{
Impact Initiatives \email{martin.barner@impact-initiatives.org}
}
