% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closure_defaults.R
\name{is_questionnaire_loaded}
\alias{is_questionnaire_loaded}
\title{Determine if a questionnarie has been loaded}
\usage{
is_questionnaire_loaded()
}
\value{
\code{TRUE} if \code{\link{load_questionnaire}} has been called successfully.
}
\description{
Checks if a questionnaire has been loaded into the global space with \code{\link{load_questionnaire}}
}
\examples{
is_questionnaire_loaded()
}
\seealso{
\code{\link{load_questionnaire}}
}
