% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{kobo_choices}
\alias{kobo_choices}
\title{Example kobo choices}
\description{
Example kobo choices
}
\references{
\url{www.impact-initiatives.org}
}
\author{
Impact Initiatives \email{martin.barner@impact-initiatives.org}
}
