% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{kobo_questions}
\alias{kobo_questions}
\title{Example kobo questions}
\description{
Example kobo questions
}
\references{
\url{www.impact-initiatives.org}
}
\author{
Impact Initiatives \email{martin.barner@impact-initiatives.org}
}
