% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closure_defaults.R
\name{question_get_choice_labels}
\alias{question_get_choice_labels}
\title{what is this?
This package is using a specific kind of closures
The factory is defined, but also the produced functions are *defined* and *documented* as part of the package.
That allows us to make functions available once data is loaded, but still provide defaults for those functions and document them with vingnettes
They consist of three parts:
1. The exported function; it calls an internal function that is not exported and therefore editable by the factory
2. A default for the internal function; usually throws and error or does some simple default behaviour
3. A factory; the factory fills the internal function with it's main code once the factory is called.}
\usage{
question_get_choice_labels(responses, variable.name)
}
\arguments{
\item{responses}{A vector of responses in kobo xml name format}

\item{variable.name}{The xml name of a kobo question. (as it appears in the kobo questionnaire and subsequently in the data column headers)
Should be a question of type 'select_one' or 'select_multiple'.}
}
\value{
A vector of strings with labels corresponding to the choices supplied in \code{responses}.
}
\description{
Here we define 1. and 2.
All of the closures here are fabricated by load_questionnaire().
Convert kobo xml choice names to labels
}
\details{
Uses a loaded kobo questionnaire to get the labels of select_one and select_multiple type questions.

To use this you must first successfully run \code{\link{load_questionnaire}}.
If conversion of a value in \code{responses} fails, the original input value is returned. This happens in the following cases:
\itemize{
  \item No questionnaire has been loaded with \code{\link{load_questionnaire}}
  \item \code{variable.name} could not be found in the questionnaire, or in the data that was supplied to \code{\link{load_questionnaire}}
  \item A value in \code{responses} is not listed in the loaded questionnaire
  \item \code{variable.name} is not listed as a question of type 'select_one' or 'select_multiple' in the questionnaire.
}

This does not work for concatenated responses of 'select_multiple' questions (e.g. strings of the form "choice_a choice_b"); The \code{responses} should have only a single response per element.
}
\examples{
question_get_choice_labels(mydata$location, "location")
}
\seealso{
\code{\link{load_questionnaire}} must be run first. The equivalent of this but for question labels is \code{\link{question_get_question_label}},
}
