% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closure_defaults.R
\name{question_get_question_label}
\alias{question_get_question_label}
\title{Convert kobo xml question names to labels}
\usage{
question_get_question_label(question.name)
}
\arguments{
\item{question.name}{The xml name of a kobo question as a string. (as it appears in the kobo questionnaire and subsequently in the data column headers)}
}
\value{
A string with the kobo label of the question / data column header
}
\description{
Uses a loaded kobo questionnaire to get the label corresponding to a question xml name / data column headers.
}
\details{
To use this you must first successfully run \code{\link{load_questionnaire}}.
If conversion to label fails, the original input value is returned. This happens in the following cases:
\itemize{
  \item No questionnaire has been loaded with \code{\link{load_questionnaire}}
  \item \code{question.name} could not be found in the questionnaire, or in the data that was supplied to \code{\link{load_questionnaire}}
}
}
\examples{
question_get_question_label("a_variable_name")
}
\seealso{
\code{\link{load_questionnaire}} must be run first. The equivalent of this but for choice labels is \code{\link{question_get_choice_labels}},
}
