% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closure_defaults.R
\name{question_in_questionnaire}
\alias{question_in_questionnaire}
\title{Determine if a question name can be found in the loaded questionnaire}
\usage{
question_in_questionnaire(question.name)
}
\arguments{
\item{question.name}{The xml name of a (potential) kobo question as a string.}
}
\value{
\code{TRUE} if the question is listed in the loaded questionnaire. \code{FALSE} otherwise.
}
\description{
Determine if a question name can be found in the loaded questionnaire
}
