% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closure_defaults.R
\name{question_is_categorical}
\alias{question_is_categorical}
\title{Determine if a kobo question is categorical}
\usage{
question_is_categorical(question.name)
}
\arguments{
\item{question.name}{The xml name of a kobo question as a string. (as it appears in the kobo questionnaire and subsequently in the data column headers)}
}
\value{
\code{TRUE} if the question is listed as a select_one or select_multiple type in the questionnaire. \code{FALSE} if the question is listed as a different type. \code{FALSE} if the question type could not be determined from the questionnaire.
}
\description{
Uses a loaded kobo questionnaire to look up the question type for a given question; returns true for select_one or select_multiple
}
\details{
To use this you must first successfully run \code{\link{load_questionnaire}}.
This does not derive the data type from any actual data; it only looks up the type defined in the questionnaire.
If type identification fails, the default return value is \code{FALSE}.This happens in the following cases:
\itemize{
  \item No questionnaire has been loaded with \code{\link[koboquest]{load_questionnaire}}
  \item \code{question.name} could not be found in the questionnaire, or in the data that was supplied to \code{\link{load_questionnaire}}
}
}
\examples{
question_is_categorical("some_numeric_kobo_xml_question_name") # FALSE
question_is_categorical("a_select_one_kobo_xml_question_name") # TRUE
question_is_categorical("a_select_multiple_kobo_xml_question_name") # TRUE
question_is_categorical("some_unidentified_string") # FALSE
}
\seealso{
\code{\link{load_questionnaire}} must be run first.
Use \code{\link{question_type}} for the most generalised way to guess the data type.
Part of the \code{question_is_*} family of functions:
testing for specific types:
\code{\link{question_is_numeric}},
\code{\link{question_is_categorical}},
\code{\link{question_is_select_one}},
\code{\link{question_is_select_multiple}},
\code{\link{question_is_sm_choice}}

parsing kobo skip-logic:
\code{\link{question_is_skipped}}
}
