% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closure_defaults.R
\name{question_is_skipped}
\alias{question_is_skipped}
\title{Determine if a variables records were skipped in the questionnaire}
\usage{
question_is_skipped(data, question.name)
}
\arguments{
\item{data}{The dataset as a data.frame in standard kobo format, with the column headers matching the question names.}

\item{question.name}{The xml name of a kobo question as a string. (as it appears in the kobo questionnaire and subsequently in the data column headers)}
}
\value{
a logical vector with one value per row in \code{data}. \code{TRUE} if a record was skipped, \code{FALSE} otherwise.
}
\description{
Uses a loaded kobo questionnaire and a dataset to determine which records have been skipped for a particular variable.
}
\details{
To use this you must first successfully run \code{\link{load_questionnaire}}. If for any reason skiplogic could not be determined, it returns \code{FALSE} for all records (with a warning).
}
\examples{
mydata<-lo
question_is_skipped("kobo_xml_question_name",mydata)
}
\seealso{
\code{\link{load_questionnaire}} should be run first.
}
