% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_questionnaire.R
\name{question_type}
\alias{question_type}
\title{question_data_type}
\usage{
question_type(variable.name, data = NULL, from.questionnaire = T,
  from.data = T)
}
\arguments{
\item{variables}{the vector or value for which the type should be determined}
}
\value{
a vector or value with variable types
}
\description{
question_data_type
}
\examples{

}
