% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closure_defaults.R
\docType{data}
\name{questionnaire}
\alias{questionnaire}
\title{what is this?
This package is using a specific kind of closures
The factory is defined, but also the produced functions are *defined* and *documented* as part of the package.
That allows us to make functions available once data is loaded, but still provide defaults for those functions and document them with vingnettes
They consist of three parts:
1. The exported function; it calls an internal function that is not exported and therefore editable by the factory
2. A default for the internal function; usually throws and error or does some simple default behaviour
3. A factory; the factory fills the internal function with it's main code once the factory is called.}
\format{An object of class \code{NULL} of length 0.}
\usage{
questionnaire
}
\description{
Here we define 1. and 2.
All of the closures here are fabricated by load_questionnaire().
}
\keyword{datasets}
