% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read.csv.auto.sep}
\alias{read.csv.auto.sep}
\title{read csv files independent from csv type; default stringsAsFactors to false}
\usage{
read.csv.auto.sep(file, stringsAsFactors = F, ...)
}
\description{
read csv files independent from csv type; default stringsAsFactors to false
}
