% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_questionnaire.R
\name{load_questionnaire}
\alias{load_questionnaire}
\title{load_questionnaire}
\usage{
load_questionnaire(data, questions, choices,
  choices.label.column.to.use = NULL)
}
\arguments{
\item{data}{data frame containing the data matching the questionnaire to be loaded.}

\item{questions}{kobo form question sheet; either as a data frame, or a single character string with the name of a csv file}

\item{choices}{questions kobo form choices sheet; either as a data frame, or a single character string with the name of a csv file}

\item{choices.label.column.to.use}{The choices table has (sometimes multiple) columns with labels. They are often called "Label::English" or similar. Here you need to provide the _name of the column_ that you want to use for labels (see example!)}
}
\value{
A list containing the original questionnaire questions and choices, the choices matched 1:1 with the data columns, and all functions created by this function relating to the specific questionnaire
}
\description{
load_questionnaire
}
\examples{

}
