% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closure_defaults.R
\name{question_get_choice_labels}
\alias{question_get_choice_labels}
\title{Convert kobo xml choice names to labels}
\usage{
question_get_choice_labels(responses, variable.name)
}
\arguments{
\item{responses}{A vector of responses in kobo xml name format}

\item{variable.name}{The xml name of a kobo question. (as it appears in the kobo questionnaire and subsequently in the data column headers)
Should be a question of type 'select_one' or 'select_multiple'.}
}
\value{
A vector of strings with labels corresponding to the choices supplied in \code{responses}.
}
\description{
Uses a loaded kobo questionnaire to get the labels of select_one and select_multiple type questions.
}
\details{
To use this you must first successfully run \code{\link{load_questionnaire}}.
If conversion of a value in \code{responses} fails, the original input value is returned. This happens in the following cases:
\itemize{
  \item No questionnaire has been loaded with \code{\link{load_questionnaire}}
  \item \code{variable.name} could not be found in the questionnaire, or in the data that was supplied to \code{\link{load_questionnaire}}
  \item A value in \code{responses} is not listed in the loaded questionnaire
  \item \code{variable.name} is not listed as a question of type 'select_one' or 'select_multiple' in the questionnaire.
}

This does not work for concatenated responses of 'select_multiple' questions (e.g. strings of the form "choice_a choice_b"); The \code{responses} should have only a single response per element.
}
\examples{
question_get_choice_labels(mydata$location, "location")
}
\seealso{
\code{\link{load_questionnaire}} must be run first. The equivalent of this but for question labels is \code{\link{question_get_question_label}},
}
