% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_questionnaire.R
\name{question_type_from_questionnaire}
\alias{question_type_from_questionnaire}
\title{variable_type}
\usage{
question_type_from_questionnaire(variables)
}
\arguments{
\item{variables}{the vector or value for which the type should be determined}
}
\value{
a vector or value with variable types
}
\description{
variable_type
}
\examples{

}
